Move exported method into package_test to avoid export_test
