--- api/client/annotations/export_test.go
@@ -1,12 +1,0 @@
-// Copyright 2022 Canonical Ltd.
-// Licensed under the AGPLv3, see LICENCE file for details.
-
-package annotations
-
-import "github.com/juju/juju/api/base"
-
-func NewClientFromCaller(caller base.FacadeCaller) *Client {
-	return &Client{
-		facade: caller,
-	}
-}
